backend/planr/pyproject.toml
[tool.poetry]
name = "planr"
version = "0.1.0"
description = ""
authors = ["cheetahbyte <bernerdoodle@outlook.de>"]

[tool.poetry.scripts]
start = "planr:start"
"db:user" = "planr.manage:Database.Proxy.user"
"db:init" = "planr.manage:Database.Proxy.init"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
fastapi-jwt = "^0.1.7"
fastapi = "^0.70.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.0"}
uvicorn = {extras = ["standart"], version = "^0.16.0"}
PyInquirer = "^1.0.3"
bcrypt = "^3.2.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
